pyproject.toml
[tool.poetry]
name = "cocloud"
version = "0.1.0"
description = "CLoud for files"
authors = ["romanlosev <losik19-94@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.42"
alembic = "^1.8.1"
isort = "^5.10.1"
asyncpg = "^0.27.0"
greenlet = "^1.1.3.post0"
orjson = "^3.8.1"
loguru = "^0.6.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^10.2.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3